Реализован функционал по добавлению растений, фото растений. Выбор фото по умолчанию для растений и клумбы для отображения в списке.

+++ app/src/main/java/com/irepka3/mygarden/domain/interactor/PlantInteractorImpl.kt
@@ -1,17 +1,21 @@
 package com.irepka3.mygarden.domain.interactor
 
 import com.irepka3.mygarden.domain.model.Plant
+import com.irepka3.mygarden.domain.repository.DirRepository
 import com.irepka3.mygarden.domain.repository.PlantRepository
+import java.lang.Exception
+import javax.inject.Inject
 
 /**
  * Реализация интерактора для работы с растениями
  *
  * Created by i.repkina on 02.11.2021.
  */
-class PlantInteractorImpl(private val plantRepository: PlantRepository): PlantInteractor {
-    override fun getPlantsAll(): List<Plant> {
-        return plantRepository.getPlantsAll()
-    }
+class PlantInteractorImpl
+@Inject constructor(
+    private val plantRepository: PlantRepository,
+    private val dirRepository: DirRepository
+) : PlantInteractor {
 
     override fun getPlantsByFlowerbed(flowerbedId: Long): List<Plant> {
         return plantRepository.getPlantsByFlowerbed(flowerbedId)
@@ -21,8 +25,8 @@ class PlantInteractorImpl(private val plantRepository: PlantRepository): PlantIn
         return plantRepository.getPlant(plantId)
     }
 
-    override fun insertPlant(plant: Plant) {
-        plantRepository.insertPlant(plant)
+    override fun insertPlant(plant: Plant): Long {
+        return plantRepository.insertPlant(plant)
     }
 
     override fun updatePlant(plant: Plant) {
@@ -30,6 +34,12 @@ class PlantInteractorImpl(private val plantRepository: PlantRepository): PlantIn
     }
 
     override fun deletePlant(plant: Plant) {
+        if (plant.plantId == null)
+            throw Exception("Invalid flowerbedId, flowerbedId is null")
+        val file = dirRepository.getPlantDir(flowerbedId = plant.flowerbedId, plantId = plant.plantId)
+        if (file.exists()) {
+            file.deleteRecursively()
+        }
         plantRepository.deletePlant(plant)
     }
 }
